+++ packages/core/src/constructors/bundle-factory.ts
@@ -1,3 +1,5 @@
+import { ShikiError } from '@shikijs/types'
+
 import type {
   Awaitable,
   BundledHighlighterOptions,
@@ -20,10 +22,9 @@ import type {
   ThemeInput,
   TokensResult,
 } from '@shikijs/types'
-
 import type { Root } from 'hast'
-import { createWasmOnigEngine } from '../engines/wasm'
-import { ShikiError } from '../error'
+
+import { createWasmOnigEngine } from '../engines/oniguruma'
 import { isSpecialLang, isSpecialTheme } from '../utils'
 import { createHighlighterCore } from './highlighter'
