+++ packages/core/src/textmate.ts
@@ -1,0 +1,1 @@
+export * from '@shikijs/vscode-textmate'
