+++ packages/core/src/types.ts
@@ -1,0 +1,1 @@
+export * from '@shikijs/types'
