--- packages/core/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "lib": ["esnext", "DOM"],
-    "rootDir": ".",
-    "module": "esnext",
-    "moduleResolution": "Bundler",
-    "resolveJsonModule": true,
-    "allowJs": true,
-    "strict": true,
-    "strictNullChecks": true,
-    "noEmit": true,
-    "preserveValueImports": false,
-    "esModuleInterop": true,
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true
-  }
-}
