+++ packages/engine-javascript/README.md
@@ -1,0 +1,9 @@
+# @shikijs/engine-javascript
+
+Engine for Shiki using JavaScript's native RegExp (experimental).
+
+[Documentation](https://shiki.style/guide/regex-engines)
+
+## License
+
+MIT
