+++ packages/engine-javascript/build.config.ts
@@ -1,0 +1,19 @@
+import { defineBuildConfig } from 'unbuild'
+
+export default defineBuildConfig({
+  entries: [
+    'src/index.ts',
+  ],
+  declaration: true,
+  rollup: {
+    emitCJS: false,
+    dts: {
+      compilerOptions: {
+        paths: {},
+      },
+    },
+  },
+  externals: [
+    'hast',
+  ],
+})
