+++ packages/engine-javascript/package.json
@@ -1,56 +1,42 @@
 {
-  "name": "@shikijs/monaco",
+  "name": "@shikijs/engine-javascript",
   "type": "module",
   "version": "1.16.3",
-  "description": "Use Shiki for Monaco Editor",
+  "description": "Engine for Shiki using JavaScript's native RegExp",
   "author": "Anthony Fu <anthonyfu117@hotmail.com>",
   "license": "MIT",
   "homepage": "https://github.com/shikijs/shiki#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/shikijs/shiki.git",
-    "directory": "packages/shiki-monaco"
+    "directory": "packages/transformers"
   },
   "bugs": "https://github.com/shikijs/shiki/issues",
   "keywords": [
     "shiki",
-    "monaco-editor"
+    "@shikijs/transformers"
   ],
   "sideEffects": false,
   "exports": {
     ".": {
       "types": "./dist/index.d.mts",
       "default": "./dist/index.mjs"
-    },
-    "./*": "./dist/*"
+    }
   },
   "main": "./dist/index.mjs",
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.mts",
-  "typesVersions": {
-    "*": {
-      "*": [
-        "./dist/*",
-        "./*"
-      ]
-    }
-  },
   "files": [
     "dist"
   ],
   "scripts": {
     "build": "unbuild",
     "dev": "unbuild --stub",
-    "prepublishOnly": "nr build",
-    "test": "vitest",
-    "play": "nr -C playground play"
+    "prepublishOnly": "nr build"
   },
   "dependencies": {
-    "@shikijs/core": "workspace:*",
     "@shikijs/types": "workspace:*",
-    "@shikijs/vscode-textmate": "catalog:"
-  },
-  "devDependencies": {
-    "monaco-editor-core": "catalog:"
+    "oniguruma-to-js": "catalog:",
+    "regex": "catalog:"
   }
 }
