+++ packages/engine-oniguruma/src/wasm-inlined.ts
@@ -1,0 +1,11 @@
+// @ts-expect-error this will be compiled to ArrayBuffer
+import binary from 'vscode-oniguruma/release/onig.wasm'
+import type { WebAssemblyInstantiator } from '@shikijs/types'
+
+export const wasmBinary = binary as ArrayBuffer
+
+export const getWasmInstance: WebAssemblyInstantiator = async (info) => {
+  return WebAssembly.instantiate(wasmBinary, info).then(wasm => wasm.instance.exports)
+}
+
+export default getWasmInstance
