+++ packages/rehype/package.json
@@ -51,6 +51,7 @@
   },
   "dependencies": {
     "@shikijs/transformers": "workspace:*",
+    "@shikijs/types": "workspace:*",
     "@types/hast": "catalog:",
     "hast-util-to-string": "catalog:",
     "shiki": "workspace:*",
