+++ packages/shiki/package.json
@@ -101,6 +101,7 @@
   },
   "dependencies": {
     "@shikijs/core": "workspace:*",
+    "@shikijs/types": "workspace:*",
     "@shikijs/vscode-textmate": "catalog:",
     "@types/hast": "catalog:"
   },
