+++ packages/twoslash/package.json
@@ -56,6 +56,7 @@
   },
   "dependencies": {
     "@shikijs/core": "workspace:*",
+    "@shikijs/types": "workspace:*",
     "twoslash": "catalog:"
   },
   "devDependencies": {
