+++ tsconfig.json
@@ -16,6 +16,9 @@
       "@shikijs/vitepress-twoslash": ["./packages/vitepress-twoslash/src/index.ts"],
       "@shikijs/markdown-it": ["./packages/markdown-it/src/index.ts"],
       "@shikijs/types": ["./packages/types/src/index.ts"],
+      "@shikijs/engine-javascript": ["./packages/engine-javascript/src/index.ts"],
+      "@shikijs/engine-oniguruma/wasm-inline": ["./packages/engine-oniguruma/src/wasm-inline.ts"],
+      "@shikijs/engine-oniguruma": ["./packages/engine-oniguruma/src/index.ts"],
       "shiki/core": ["./packages/shiki/src/core.ts"],
       "shiki/wasm": ["./packages/shiki/src/wasm.ts"],
       "shiki/langs": ["./packages/shiki/src/langs.ts"],
